requirements.txt
Django==3.2.5
djangorestframework==3.12.4
django-cors-headers==3.5.0
pika==1.1.0
PyJWT
requests
mysqlclient==2.1.0
django-prometheus==2.1.0